requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
paho-mqtt==1.5.1
requests==2.26.0
urllib3==1.26.6
